+++ src/numerics/spline/cl_Spline.hpp
@@ -26,8 +26,8 @@ namespace belfem
     {
         void
        create_helpmatrix(
-              const real & aSize,
-              const real & aDeltaX,
+              const index_t aSize,
+              const real    aDeltaX,
               SpMatrix & aA,
               const SplineBC aStartBC = SplineBC::NoCurvature,
               const SplineBC aEndBC = SplineBC::NoCurvature );